# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"preferences": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiErrorBody": {
				"properties": {
					"errorCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"errorCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiFailedResponse": {
				"description": "전역 에러 핸들러(res.error)가 내려주는 실패 응답 형식",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"$ref": "#/components/schemas/ApiErrorBody"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"user_name": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"user_phone": {
						"type": "string"
					},
					"user_gender": {
						"type": "string",
						"enum": [
							"여성",
							"남성"
						]
					},
					"birth_data": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"일반 사용자",
							"가게 운영자"
						]
					},
					"point": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"user_name",
					"nickname",
					"user_phone",
					"user_gender",
					"birth_data",
					"address",
					"role",
					"point",
					"email",
					"preferences",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playground-umc-7th-nodejs",
		"version": "1.0.0",
		"description": "UMC 테스트 프로젝트",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "user@example.com",
											"user_name": "홍길동",
											"nickname": "UMC10th",
											"user_phone": "01012345678",
											"user_gender": "여성",
											"birth_data": "2000-01-01",
											"address": "서울시",
											"role": "일반 사용자",
											"point": 0,
											"password": "password123",
											"preferences": [
												1
											]
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "요청 형식 오류 (필수 필드 누락·잘못된 필드명·존재하지 않는 선호 카테고리 ID 등)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailedResponse"
								}
							}
						}
					},
					"409": {
						"description": "중복된 이메일 (U001)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailedResponse"
								}
							}
						}
					},
					"500": {
						"description": "서버 오류 (COMMON500)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailedResponse"
								}
							}
						}
					}
				},
				"description": "회원가입 API",
				"summary": "회원가입을 처리하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
...
		"/users/mypage": {
			"get": {
				"operationId": "HandleMypage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요 (미들웨어)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailedResponse"
								}
							}
						}
					}
				},
				"description": "마이페이지 (HTML, 로그인 필요)",
				"summary": "쿠키 기반 로그인이 필요한 페이지입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}